chore: Move disconnect repo endpoint to endpoints section of app
