+++ webapp/endpoints/publisher/builds.py
@@ -1,3 +1,6 @@
+# Standard library
+import os
+
 # Packages
 import flask
 from canonicalwebteam.store_api.dashboard import Dashboard
@@ -8,6 +11,8 @@ from requests.exceptions import HTTPError
 from webapp.helpers import api_publisher_session, launchpad
 from webapp.api.github import GitHub, InvalidYAML
 from webapp.decorators import login_required
+
+GITHUB_WEBHOOK_HOST_URL = os.getenv("GITHUB_WEBHOOK_HOST_URL")
 
 dashboard = Dashboard(api_publisher_session)
 
@@ -123,3 +128,37 @@ def post_build(snap_name):
         )
 
     return flask.jsonify({"success": True, "build_id": build_id})
+
+
+@login_required
+def post_disconnect_repo(snap_name):
+    details = dashboard.get_snap_info(flask.session, snap_name)
+
+    lp_snap = launchpad.get_snap_by_store_name(snap_name)
+    launchpad.delete_snap(details["snap_name"])
+
+    # Try to remove the GitHub webhook if possible
+    if flask.session.get("github_auth_secret"):
+        github = GitHub(flask.session.get("github_auth_secret"))
+
+        try:
+            gh_owner, gh_repo = lp_snap["git_repository_url"][19:].split("/")
+
+            old_hook = github.get_hook_by_url(
+                gh_owner,
+                gh_repo,
+                f"{GITHUB_WEBHOOK_HOST_URL}api/{snap_name}/webhook/notify",
+            )
+
+            if old_hook:
+                github.remove_hook(
+                    gh_owner,
+                    gh_repo,
+                    old_hook["id"],
+                )
+        except HTTPError:
+            pass
+
+    return flask.redirect(
+        flask.url_for(".get_snap_builds", snap_name=snap_name)
+    )
